renamed game folder to components in client folder for ease of use
